+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
++ **[ENHANCEMENT]** Libs: Change namespace name to STM32.
+
 ## [v1.0.0] - 08-12-2025
 
 + Initial release.